Rename deserialize.ts to serialize.ts.
